Move the LaserPositionUpdate tracking hooks into Ext/LaserDraw/Hooks.cpp
Pure relocation - the implementation is copied verbatim from
src/Misc/Hooks.LaserDraw.cpp, which keeps only the laser drawing and
Z-adjust hooks.
